+++ CHANGELOG.md
@@ -12,6 +12,7 @@ See [Releases](https://gitlab.com/gitlab-org/step-runner/-/releases) for when ch
 - Publishing steps as an OCI image publishes the `latest` tag if publishing
   latest version. See !205. 
 - The server command accepts a unix socket path on which to listen. See !206
+- Authenticate to OCI registries using the Docker config file. See !207.
 
 ## v0.7.0
 - Inline script step to reduce external network calls. See !186.